+++ bamboo/query.py
@@ -1,9 +1,9 @@
-"""Helper classes for creating nested and inverted queries."""
+"""Classes representing elasticsearch queries."""
 from abc import ABCMeta, abstractproperty
 from collections import defaultdict
 from copy import deepcopy
 
-from .exceptions import BadOperatorError
+from .utils import dict_to_params
 
 
 def boost(query, value):
@@ -48,22 +48,31 @@ class Query(object):
     def key(self):
         """Get type of elasticsearch query object."""
 
+    def __repr__(self):
+        return '{}(field={}, value={}, boost={})'.format(
+            type(self).__name__, self.field, self.value, self._boost
+        )
+
     def __and__(self, other):
         if other is None:
             return self
-        if not isinstance(other, Query):
+        if isinstance(other, Bool):
             return NotImplemented
         return Bool(must=[self, other])
+    __rand__ = __and__
 
     def __or__(self, other):
-        if not isinstance(other, Query):
+        if other is None:
+            return self
+        if isinstance(other, Bool):
             return NotImplemented
         return Bool(should=[self, other])
+    __ror__ = __or__
 
     def __invert__(self):
-        return Bool(must_not=[self])
+        return Bool(must_not=self)
 
-    def body(self):
+    def __call__(self):
         """Return the query body formatted for elasticsearch."""
         if self._boost:
             return self._boosted_query
@@ -77,6 +86,7 @@ class Query(object):
             }
         }
 
+    @property
     def _boosted_query(self):
         return {
             self.key: {
@@ -88,18 +98,52 @@ class Query(object):
         }
 
     def boost(self, value):
-        """Boost the weight of the query by the value."""
+        """Boost the weight of the query by the value.
+
+        Returns:
+            Query: Query with boost applied.
+        """
         new = deepcopy(self)
         new._boost = value
         return new
 
-    __rand__ = __and__
-
 
 class Term(Query):
-    """Find documents containing the exact term specified in the inverted index."""
+    """Find documents containing the exact term specified in the inverted index.
+
+    The field is not analyzed.
+    """
 
     key = 'term'
+
+
+class Terms(Query):
+    """Find documents containing any of the provided terms.
+
+    The field is not analyzed.
+    """
+
+    key = 'terms'
+
+    def __init__(self, field, value, boost=None):
+        """Init Query.
+
+        Args:
+            field (str): Field the query is conditioned upon
+            value (list): List of terms to match
+            boost (float, optional): Weight given to this query.
+                Default None.
+        """
+        super(Terms, self).__init__(field, value, boost)
+
+    @property
+    def _boosted_query(self):
+        return {
+            self.key: {
+                self.field: self.value,
+                'boost': self._boost
+            }
+        }
 
 
 class Regexp(Query):
@@ -165,13 +209,17 @@ class Exists(Query):
         """
         super(Exists, self).__init__('field', field, boost)
 
+    @property
     def _boosted_query(self):
         return {
             self.key: {
-                'field': self.field,
+                self.field: self.value,
                 'boost': self._boost
             }
         }
+
+    def __repr__(self):
+        return 'Exists(field={}, boost={})'.format(self.value, self._boost)
 
 
 class Range(Query):
@@ -179,33 +227,57 @@ class Range(Query):
 
     key = 'range'
 
-    def __init__(self, field, value, boost=None):
+    def __init__(self, field, boost=None):
         """Init Query.
 
         Args:
             field (str): Field the query is conditioned upon
-            value (obj): Value the condition is applying
             boost (float, optional): Weight given to this query.
                 Default None.
         """
-        super(Range, self).__init__(self, field, value, boost)
+        self.field = field
+        self._boost = boost
         self._operators = {}
 
     def greater_than(self, value):
-        """Apply greater than operator to value."""
-        self._operators['gt'] = value
+        """Apply greater than operator to value.
+
+        Returns:
+            Range: Range query with greater than value applied.
+        """
+        new = deepcopy(self)
+        new._operators['gt'] = value
+        return new
 
     def greater_than_or_equal(self, value):
-        """Apply greater than or equal operator to value."""
-        self._operators['gte'] = value
+        """Apply greater than or equal operator to value.
+
+        Returns:
+            Range: Range query with greater than or equal value applied.
+        """
+        new = deepcopy(self)
+        new._operators['gte'] = value
+        return new
 
     def less_than(self, value):
-        """Apply less than operator to value."""
-        self._operators['lt'] = value
+        """Apply less than operator to value.
+
+        Returns:
+            Range: Range query with less than value applied.
+        """
+        new = deepcopy(self)
+        new._operators['lt'] = value
+        return new
 
     def less_than_or_equal(self, value):
-        """Apply less than or equal operator to value."""
-        self._operators['lte'] = value
+        """Apply less than or equal operator to value.
+
+        Returns:
+            Range: Range query with less than or equal value applied.
+        """
+        new = deepcopy(self)
+        new._operators['lte'] = value
+        return new
 
     @property
     def _query(self):
@@ -223,19 +295,21 @@ class Range(Query):
             }
         }
 
+    def __repr__(self):
+        return 'Range({})'.format(dict_to_params(self._operators))
+
 
 class Bool(Query):
-    """Class for keeping track of complex and nested queries.
+    """Find documents matching boolean combinations of other queries.
 
     Attributes:
         body: Query body formatted for elasticsearch
     """
 
-    _valid_params = {'must', 'must_not', 'should'}
-
     key = 'bool'
+    _validation_msg = 'At least one initialization parameter must be supplied.'
 
-    def __init__(self, **params):
+    def __init__(self, must=None, must_not=None, should=None, boost=None):
         """Init Bool.
 
         Args:
@@ -243,77 +317,99 @@ class Bool(Query):
             must_not (list): A list of queries that must not exist
             should (list): A list of queries that do not have to exist
                 but affect scoring/sorting
+            boost (float, optional): Weight given to this query.
+                Default None.
         """
-        # must, must_not and should (and filter?) as defined attributes
-        # validate at least one param is supplied
-        # validate_params should allow set and if set then convert to
-        # List[set]
-        params = dict(self.__validate_params(params))
-        self.update(params)
+        # TODO: add filter param
+        self.params = {
+            'must': must or [],
+            'must_not': must_not or [],
+            'should': should or []
+        }
+        self._boost = boost
+        self.__validate_params()
 
-    @classmethod
-    def __validate_params(cls, params):
+    def __repr__(self):
+        return 'Bool({})'.format(dict_to_params(self.params))
+
+    def __validate_params(self):
         """Ensure parameters are in correct format."""
-        assert params, "Params are empty"
-        for k, v in params.items():
-            assert k in cls._valid_params, "Bad parameter"
-            if v:
-                assert isinstance(v, list), "Parameter must be a list"
-                yield k, filter(None, v)
+        assert any(self.params.values()), self._validation_msg
+        for key, param in self.params.items():
+            if isinstance(param, Query):
+                self.params[key] = [param]
 
-    @staticmethod
-    def _validate_obj(item):
-        """Ensure inputs are of correct type."""
-        if not isinstance(item, Query):
-            raise BadOperatorError(item)
+    @property
+    def must(self):
+        """Get the must conditions."""
+        return self.params['must']
+
+    @property
+    def must_not(self):
+        """Get the must_not conditions."""
+        return self.params['must_not']
+
+    @property
+    def should(self):
+        """Get the should conditions."""
+        return self.params['should']
 
     def __add__(self, other):
-        params = self.merge_params(self, other)
+        params = self.merge_params(self.params, other.params)
         return Bool(**params)
 
     def __and__(self, other):
-        if other is None:
-            return self
-        self._validate_obj(other)
-        if isinstance(other, Query):
-            return self + Bool(must=[other])
-        if isinstance(self.body, Query):
-            return Bool(must=[self]) + other
-        if not set(self) & set(other):
+        if not isinstance(other, Bool):
+            return self + Bool(must=other)
+
+        # reduce must(must()) to must() when dataframe is chained
+        if not set(self.filtered_params) & set(other.filtered_params):
             return self + other
-        return Bool(must=[self.body, other.body])
+
+        return Bool(must=[self, other])
 
     def __or__(self, other):
-        self._validate_obj(other)
-        if isinstance(other, Query):
-            return self + Bool(should=[other])
-        return Bool(
-            should=[self.body, other.body]
-        )
+        if not isinstance(other, Bool):
+            return self + Bool(should=other)
+        return Bool(should=[self, other])
 
     def __invert__(self):
-        must_not = self.get('must_not', [])
-        must = self.get('must', [])
-        should = self.get('should', [])
         return Bool(
-            must=[self.__negate(i) for i in should] + must_not,
-            should=[self.__negate(i) for i in must]
+            must=[self.__negate(i) for i in self.should] + self.must_not,
+            should=[self.__negate(i) for i in self.must]
         )
 
-    def __negate(self, value):
+    def __negate(self, query):
         """Convert `must` queries to `must_not`."""
-        must_not = value.get('bool', {}).get('must_not')
-        if must_not:
-            return Bool(must=must_not).body
-        return Bool(must_not=[value]).body
+        if isinstance(query, Bool):
+            return ~query
+        return Bool(must_not=query)
 
     @property
-    def body(self):
-        """Query body formatted for elasticsearch."""
-        must = self.get('must', [])
-        if len(self) == 1 == len(must):
-            return Query(must[0])
-        return {'bool': self}
+    def _query(self):
+        params = self._finalize_params()
+        # if must is only parameter and contains only one condition
+        # then convert that one condition to a stand-alone query
+        if len(params.get('must', [])) == 1 == len(params):
+            return params['must'][0]
+        return {self.key: params}
+
+    @property
+    def _boosted_query(self):
+        params = self._finalize_params()
+        return {self.key: dict(params, **{'boost': self._boost})}
+
+    @property
+    def filtered_params(self):
+        """Get only the parameters populated with conditions."""
+        return {k: v for k, v in self.params.items() if v}
+
+    def _finalize_params(self):
+        return {
+            key: [param() for param in params]
+            for key, params in self.params.items()
+            if params
+        }
 
     @staticmethod
     def merge_params(a, b):
@@ -330,7 +426,3 @@ class Bool(Query):
         for k, v in b.items():
             d[k].extend(v)
         return d
-
-    __rand__ = __and__
-    __radd__ = __add__
-    __ror__ = __or__